refactored custom button with CSS module and react styles passed in as props
