+++ src/css/CustomButton.module.css
@@ -1,0 +1,23 @@
+.customButton {
+  border: 2px solid #0C666E; 
+  border-radius: 5px;
+  background-color: #0C666E;
+  color: white;
+  /* height: 4rem;
+  width: 24rem; */
+  font-family: 'Monserrat', sans-serif;
+  font-size: 1.5em;
+}
+
+.customButton:hover {
+  background-color: transparent;
+  color: #0C666E;
+  text-decoration: underline;
+}
+
+@media screen and (max-width: 320px) {
+  .customButton {
+    height: 1em;
+    width: 5em;
+  }
+}
