+++ src/css/SplashImage.module.css
@@ -4,7 +4,7 @@
   margin-left: 2rem;
 }
 
-@media screen and  (max-width: 600px) {
+@media screen and (max-width: 600px) {
   .splashImage {
     display: none;
   }
