+++ src/views/containers/SplashContainer.tsx
@@ -4,13 +4,13 @@ import { Grid } from '@material-ui/core';
 import Header from '../components/multi/Header';
 import SplashText from '../components/splash/SplashText';
 import SplashImage from '../components/splash/SplashImage';
-import CustomButton from '../components/multi/CustomButton';
+import CustomButton from '../components/splash/CustomButton';
 import Footer from '../components/splash/Footer';
 
 import styles from '../../css/SplashContainer.module.css';
 
 const SplashContainer = () => {
-  //redirect to company search on click
+  //button handleClick: redirect to company search 
   let history = useHistory();
   const handleClick = () => {
     history.push('/company-search');
@@ -29,16 +29,13 @@ const SplashContainer = () => {
           <SplashText />
           <div className={styles.splashButtonContainer}>
             <CustomButton
+              className="splash-button"
               buttonText="Discover your next investment"
               onClick={handleClick}
-              backgroundColor="#0C666E"
-              padding="10px"
-              fontSize="25px"
-              border="2px solid"
-              borderColor="#0C666E"
-              borderRadius="5px"
-              color="white"
-              margin="50px 0 0 0"
+              height="4rem"
+              width="24rem"
+              margin="1rem 0 0 0"
+              padding="0"
             />
           </div>
         </Grid>
